requirements.txt
Appium-Python-Client==4.0.0
selenium==4.23.1
python-dotenv==1.0.1
